feature: align delete button behavior across pages - delete button will open a dialog asking confirmation everywhere

+++ client/src/components/Entries.tsx
@@ -63,6 +63,7 @@ export class Entries extends React.PureComponent<EntriesProps, EntriesState> {
   }
 
   onEntryDelete = async (entryId: string) => {
+    if (!window.confirm('Wirklich löschen?')) return
     try {
       await deleteEntry(this.props.auth.getIdToken(), entryId)
       this.setState({
